Create Challenge 2: Box office hits database
This database contains an incomplete list of box office hits and their release year. In this challenge, I'm going to get the results back out of the database in different ways!
In step 1, I selected all the movies.
In step 2, I added a second query after the first, that retrieves only the movies that were released in the year 2000 or later, not before. Then I sorted the results so that the earlier movies are listed first. I had 2 SELECT statements after this step.