Create Challenge 3: TODO list database stats
Here's a table containing a TODO list with the number of minutes it will take to complete each item. I inserted another item to my todo list with the estimated minutes it will take.
Then I selected the SUM of minutes it will take to do all of the items on my TODO list. I only had one SELECT statement.