Create Challenge 5: Playlist maker
I've created a database of songs and artists, and I made playlists from them in this challenge. In the first step, I selected the title of all the songs by the artist named 'Queen'.
Then I made a 'Pop' playlist. In preparation, I selected the name of all of the artists from the 'Pop' genre. (Tip: Make sure I type it 'Pop', SQL considers that different from 'pop'.)
To finish creating the 'Pop' playlist, I added another query that will select the title of all the songs from the 'Pop' artists. It used IN on a nested subquery that's based on my previous query.