Create Challenge 7: Gradebook
I've created a database to track student grades, with their name, number grade, and what percent of activities they've completed. In the first step, I selected all of the rows, and display the name, number_grade, and percent_completed, which I can compute by multiplying and rounding the fraction_completed column.
Then, step 2 is a little tricky. The goal is a table that shows how many students have earned which letter_grade. I outputed the letter_grade by using CASE with the number_grade column, outputting 'A' for grades > 90, 'B' for grades > 80, 'C' for grades > 70, and 'F' otherwise. Then I can use COUNT with GROUP BY to show the number of students with each of those grades.