Rename Challenge 8: Customer's orders to Challenge 9: Customer's orders
I've created a database for customers and their orders. Not all of the customers have made orders, however. I came up with a query that lists the name and email of every customer followed by the item and price of orders they've made. I used a LEFT OUTER JOIN so that a customer is listed even if they've made no orders, and don't add any ORDER BY.
Then I created another query that will result in one row per each customer, with their name, email, and total amount of money they've spent on orders. I sorted the rows according to the total money spent, from the most spent to the least spent.
(Tip: You should always GROUP BY on the column that is most likely to be unique in a row.)