Create Challenge 9: Sequels in SQL
I've created a table with all the 'Harry Potter' movies, with a sequel_id column that matches the id of the sequel for each movie. I issued a SELECT that will show the title of each movie next to its sequel's title (or NULL if it doesn't have a sequel).