Challenge 07: Gradebook
In this challenge, I will be working with a database designed to track student grades. The database contains information such as the student's name, their numerical grade, and the percentage of activities they have completed. My objective is to analyze this data and perform two key steps.

In the first step, I retrieved all the rows from the database and display the student's name, numerical grade, and the percentage of activities completed. To calculate the percentage completed, I multiplied the fractional value stored in the "fraction_completed" column by 100 and round it to the nearest whole number. This allowed me to obtain a clear view of each student's progress in completing activities.

Moving on to step 2, I encountered a more challenging task. My goal was to create a table that shows the distribution of letter grades earned by the students. To achieve this, I utilized the CASE statement with the "number_grade" column. I assigned the corresponding letter grade based on specific conditions: 'A' for grades above 90, 'B' for grades above 80, 'C' for grades above 70, and 'F' for any other grade. By using the COUNT function in conjunction with the GROUP BY clause, I can determine the number of students who have earned each letter grade.

By completing these steps, I gained insights into the students' grades, their completion rates, and the distribution of letter grades within the class. Let's dive into the data and analyze the performance of the students in this gradebook challenge.