Rename Challenge 8: Bobby's Hobbies to Challenge 08: Bobby's Hobbies
In this challenge, I explore a database that tracks people and their hobbies. The database consists of two tables: "persons" and "hobbies." Each row in the "hobbies" table is related to a corresponding row in the "persons" table through the "person_id" column. My task involves multiple steps to manipulate and extract relevant information from the database.

In the first step, I begin by inserting a new row in the "persons" table, representing a person named "Bobby McBobbyFace" with an age of 12. Following that, I insert a corresponding row in the "hobbies" table, associating the newly inserted person with their hobby.

Moving on to step 2, I perform a join between the "persons" and "hobbies" tables. This allows me to retrieve the data in a way that displays each person's name alongside their respective hobby. By matching the "person_id" column in the "hobbies" table with the "id" column in the "persons" table, I can establish the relationship between the two tables.

In step 3, an additional query is introduced to specifically display the name and hobbies of a particular person, namely 'Bobby McBobbyFace'. This is accomplished by combining the JOIN operation with the WHERE clause, which filters the results to only include rows where the person's name is 'Bobby McBobbyFace'.

Through these steps, I can gain insights into the hobbies of various individuals and specifically focus on the hobbies associated with 'Bobby McBobbyFace'. Let's proceed further and explore the fascinating world of Bobby's hobbies within this database.