Challenge 09: Customer's orders
In this challenge, I have created a database that stores information about customers and their orders. However, not all customers have made orders yet.

To start off, I have a query that lists the name and email of every customer, along with the item and price of the orders they have made. To ensure that even customers with no orders are included in the result, I have used a LEFT OUTER JOIN. Additionally, I haven't specified any ORDER BY in this query.

Moving on, I have another query that aims to provide one row per customer, displaying their name, email, and the total amount of money they have spent on orders. To achieve this, I have used the SUM function to calculate the total money spent and sorted the rows based on the total money spent, from the highest to the lowest.

Please note that it's important to use the GROUP BY clause on the column that is most likely to be unique within a row. In this case, I have grouped the results by the customer's name.

To give you some context, I have also included the SQL code used to create the tables for customers and orders, as well as the INSERT statements to populate them with sample data.

Feel free to explore and analyze the results obtained from these queries. Happy coding!