Create Challenge 13: Clothing alterations
In this challenge, I will be working with a database that contains information about different clothes. My objective is to make specific modifications to the table structure and populate it with accurate data.

To begin, I have a table named "clothes" with columns for ID, type, and design. However, I realize that I also need a "price" column to represent the cost of each item. I will use the ALTER statement to add the "price" column to the existing table structure. Once the alteration is complete, I will execute a SELECT statement to view all the columns in each row, providing me with a comprehensive overview of the updated table.

Next, I will assign a price to each item in the table using the UPDATE statement. Item 1 will be assigned a price of 10 dollars, item 2 will be priced at 20 dollars, and item 3 will be set at 30 dollars. To verify the success of the updates, I will perform another SELECT statement to display all the rows and ensure that the prices have been assigned correctly.

Afterward, I will insert a new item into the table that includes all three attributes, including the "price" column. By executing an INSERT INTO statement, I will add this new item to the table. To confirm the success of the insertion, I will perform a final SELECT statement to display all the rows in the "clothes" table.

By executing the provided SQL statements, I can successfully modify the table structure, assign prices to each item, insert a new item with all attributes, and verify the changes using SELECT statements.