Project 04: App impersonator
In this project, I will explore the concept of simulating app functionality using SQL statements. I will imagine a favorite app that stores data, such as a game that keeps track of scores or a social media app that allows users to post updates. For the purpose of this project, I will assume that the app stores its data in a SQL database.

Let's consider an example of a social media app called "Facebook." In Facebook, users can post updates to share their thoughts, experiences, and more. I will create a table to store the posts, insert some example rows, and then use SQL statements to emulate common app functionalities like editing and deleting data.

To begin, I will create a table named "posts" to store the data. The table will have columns such as post_id, user_id, and post_content to represent the unique post identifier, user identifier, and the content of the post, respectively.

After creating the table, I will insert a few example rows to populate it with data. These rows will simulate posts made by different users in the Facebook app, each with a unique post_id, user_id, and post_content.

Next, I will use the UPDATE statement to demonstrate what happens when I edit data in the app. I will modify the post_content of a specific post, emulating the action of updating a post in the Facebook app.

Furthermore, I will utilize the DELETE statement to imitate the process of deleting data in the app. I will remove a particular post from the "posts" table, simulating the action of deleting a post in the Facebook app.

To observe the results of our SQL statements, I will execute SELECT statements to retrieve the contents of the "posts" table at different stages of our simulation. This will allow me to verify the changes made through the UPDATE and DELETE operations.

By completing this project, I will gain a better understanding of how app functionalities can be simulated using SQL statements and how data can be manipulated within a SQL database to reflect the actions performed in a real app.

Now, let's dive into the code and explore the SQL statements involved in this app impersonator project.